Move stream functions to stream.c file.
