+++ picoquic/frames.c
@@ -1540,115 +1540,6 @@ const uint8_t* picoquic_decode_stream_frame(picoquic_cnx_t* cnx, const uint8_t* 
     return bytes;
 }
 
-picoquic_stream_head_t* picoquic_find_ready_stream_path(picoquic_cnx_t* cnx, picoquic_path_t * path_x, int is_coalesced)
-{
-    picoquic_stream_head_t* first_stream = cnx->first_output_stream;
-    picoquic_stream_head_t* stream = first_stream;
-    picoquic_stream_head_t* found_stream = NULL;
-
-
-    /* Look for a ready stream */
-    while (stream != NULL) {
-        int has_data = 0;
-        picoquic_stream_head_t* next_stream = stream->next_output_stream;
-
-        if (next_stream != NULL && is_coalesced && next_stream->is_not_coalesced) {
-            stream = next_stream->next_output_stream;
-            continue;
-        }
-
-        if (found_stream != NULL && stream->stream_priority > found_stream->stream_priority) {
-            /* All the streams at that priority level have been examined,
-             * the current selection is validated */
-            break;
-        }
-
-        /* The tests for "have data" should excatly replicate the tests in
-         * the formating of a stream frame */
-        if (stream->stop_sending_requested && !stream->stop_sending_sent) {
-            /* will send a stop sending frame.
-            * this takes precedence over FIFO vs round-robin processing */
-            found_stream = stream;
-            has_data = 1;
-            break;
-        }
-        else if (stream->reset_sent) {
-            /* No data will be sent after a reset */
-            has_data = 0;
-        }
-        else if (stream->reset_requested && 
-            (stream->reliable_size == 0 || picoquic_check_sack_list(&stream->sack_list, 0, stream->reliable_size))) {
-            /* will queue a reset frame --
-            * this takes precedence over FIFO vs round-robin processing */
-            found_stream = stream;
-            has_data = 1;
-            break;
-        }
-        else if (cnx->maxdata_remote > cnx->data_sent && stream->sent_offset < stream->maxdata_remote && (stream->is_active ||
-            (stream->send_queue != NULL && stream->send_queue->length > stream->send_queue->offset) ||
-            (stream->fin_requested && !stream->fin_sent))) {
-            has_data = 1;
-        }
-        else {
-            has_data = 0;
-        }
-
-        /* implement affinity scheduling */
-        if (has_data && path_x != NULL && stream->affinity_path != path_x && stream->affinity_path != NULL) {
-            /* Only consider the streams that meet path affinity requirements */
-            has_data = 0;
-        }
-        
-        if (has_data) {
-            /* Check that this stream is actually available for sending data */
-            if (stream->sent_offset == 0) {
-                if (IS_CLIENT_STREAM_ID(stream->stream_id) == cnx->client_mode) {
-                    if (stream->stream_id > ((IS_BIDIR_STREAM_ID(stream->stream_id)) ? cnx->max_stream_id_bidir_remote : cnx->max_stream_id_unidir_remote)) {
-                        has_data = 0;
-                    }
-                }
-            }
-            if (has_data) {
-                /* Something can be sent */
-                if ((stream->stream_priority & 1) != 0) {
-                    /* This priority level requests FIFO processing, so we return the first available stream */
-                    found_stream = stream;
-                    break;
-                }
-                else if (found_stream == NULL || stream->last_time_data_sent < found_stream->last_time_data_sent) {
-                    /* Select this stream, but need to check if another stream should go before in round robin order */
-                    found_stream = stream;
-                }
-            }
-        }
-        else if (((stream->fin_requested && stream->fin_sent) || (stream->reset_requested && stream->reset_sent)) && (!stream->stop_sending_requested || stream->stop_sending_sent)) {
-            /* If stream is exhausted, remove from output list */
-            picoquic_remove_output_stream(cnx, stream);
-
-            picoquic_delete_stream_if_closed(cnx, stream);
-        }
-        else {
-            if (stream->is_active ||
-                (stream->send_queue != NULL && stream->send_queue->length > stream->send_queue->offset)) {
-                if (stream->sent_offset >= stream->maxdata_remote) {
-                    cnx->stream_blocked = 1;
-                }
-                else if (cnx->maxdata_remote <= cnx->data_sent) {
-                    cnx->flow_blocked = 1;
-                }
-            }
-        }
-        stream = next_stream;
-    }
-
-    return found_stream;
-}
-
-picoquic_stream_head_t* picoquic_find_ready_stream(picoquic_cnx_t* cnx)
-{
-    return picoquic_find_ready_stream_path(cnx, NULL, 0);
-}
-
 /* Management of BLOCKED signals
  */
