+++ picoquic/picoquic.vcxproj.filters
@@ -171,6 +171,9 @@
     <ClCompile Include="qmux.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="streams.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="picoquic.h">
