+++ picoquic/quicctx.c
@@ -3428,427 +3428,6 @@ int picoquic_renew_connection_id(picoquic_cnx_t* cnx, int path_id)
     return ret;
 }
 
-/* stream data splay management */
-int64_t picoquic_stream_data_node_compare(void* l, void* r)
-{
-    /* Offset values are from 0 to 2^62-1, which means we are not worried with rollover */
-    return ((picoquic_stream_data_node_t*)l)->offset - ((picoquic_stream_data_node_t*)r)->offset;
-}
-
-picosplay_node_t* picoquic_stream_data_node_create(void* value)
-{
-    return &((picoquic_stream_data_node_t*)value)->stream_data_node;
-}
-
-
-void* picoquic_stream_data_node_value(picosplay_node_t* node)
-{
-    return (void*)((char*)node - offsetof(struct st_picoquic_stream_data_node_t, stream_data_node));
-}
-
-void picoquic_stream_data_node_recycle(picoquic_stream_data_node_t* stream_data)
-{
-    if (stream_data->quic->nb_data_nodes_in_pool < PICOQUIC_MAX_PACKETS_IN_POOL) {
-        stream_data->next_stream_data = stream_data->quic->p_first_data_node;
-        stream_data->quic->p_first_data_node = stream_data;
-        stream_data->quic->nb_data_nodes_in_pool++;
-    }
-    else {
-        stream_data->quic->nb_data_nodes_allocated--;
-        free(stream_data);
-    }
-}
-
-void picoquic_stream_data_node_delete(void* UNUSED(tree), picosplay_node_t* node)
-{
-    picoquic_stream_data_node_t* stream_data = (picoquic_stream_data_node_t*)picoquic_stream_data_node_value(node);
-
-    picoquic_stream_data_node_recycle(stream_data);
-}
-
-picoquic_stream_data_node_t* picoquic_stream_data_node_alloc(picoquic_quic_t* quic)
-{
-    picoquic_stream_data_node_t* stream_data = quic->p_first_data_node;
-    
-    if (stream_data == NULL) {
-        stream_data = (picoquic_stream_data_node_t*)
-            malloc(sizeof(picoquic_stream_data_node_t));
-
-        if (stream_data != NULL) {
-            /* It might be sufficient to zero the metadata, but zeroing everything
-             * appears safer, and does not confuse checkers like valgrind.
-             */
-            memset(stream_data, 0, sizeof(picoquic_stream_data_node_t));
-            stream_data->quic = quic;
-            quic->nb_data_nodes_allocated++;
-            if (quic->nb_data_nodes_allocated > quic->nb_data_nodes_allocated_max) {
-                quic->nb_data_nodes_allocated_max = quic->nb_data_nodes_allocated;
-            }
-        }
-    }
-    else {
-        quic->p_first_data_node = stream_data->next_stream_data;
-        stream_data->next_stream_data = NULL;
-        stream_data->bytes = NULL;
-        quic->nb_data_nodes_in_pool--;
-    }
-
-    return stream_data;
-}
-
-
-/* Stream splay management */
-
-static int64_t picoquic_stream_node_compare(void *l, void *r)
-{
-    /* STream values are from 0 to 2^62-1, which means we are not worried with rollover */
-    return ((picoquic_stream_head_t*)l)->stream_id - ((picoquic_stream_head_t*)r)->stream_id;
-}
-
-static picosplay_node_t * picoquic_stream_node_create(void * value)
-{
-    return &((picoquic_stream_head_t *)value)->stream_node;
-}
-
-
-static void * picoquic_stream_node_value(picosplay_node_t * node)
-{
-    return (void*)((char*)node - offsetof(struct st_picoquic_stream_head_t, stream_node));
-}
-
-void picoquic_clear_stream(picoquic_stream_head_t* stream)
-{
-    picoquic_stream_queue_node_t* ready = stream->send_queue;
-    picoquic_stream_queue_node_t* next;
-
-    while ((next = ready) != NULL) {
-        ready = next->next_stream_data;
-        if (next->bytes != NULL) {
-            free(next->bytes);
-        }
-        free(next);
-    }
-    stream->send_queue = NULL;
-    if (stream->is_output_stream) {
-        picoquic_remove_output_stream(stream->cnx, stream);
-    }
-    picosplay_empty_tree(&stream->stream_data_tree);
-    picoquic_sack_list_free(&stream->sack_list);
-}
-
-
-static void picoquic_stream_node_delete(void * UNUSED(tree), picosplay_node_t * node)
-{
-    picoquic_stream_head_t * stream = picoquic_stream_node_value(node);
-
-    picoquic_clear_stream(stream);
-
-    free(stream);
-}
-
-/* Management of streams */
-
-picoquic_stream_head_t * picoquic_stream_from_node(picosplay_node_t * node)
-{
-#ifdef TOO_CAUTIOUS
-    return(picoquic_stream_head_t *)((node == NULL)?NULL:picoquic_stream_node_value(node));
-#else
-    return (picoquic_stream_head_t *)node;
-#endif
-}
-
-picoquic_stream_head_t * picoquic_first_stream(picoquic_cnx_t* cnx)
-{
-#ifdef TOO_CAUTIOUS
-    return picoquic_stream_from_node(picosplay_first(&cnx->stream_tree));
-#else
-    return (picoquic_stream_head_t *)picosplay_first(&cnx->stream_tree);
-#endif
-}
-
-picoquic_stream_head_t * picoquic_last_stream(picoquic_cnx_t* cnx)
-{
-#ifdef TOO_CAUTIOUS
-    return picoquic_stream_from_node(picosplay_last(&cnx->stream_tree));
-#else
-    return (picoquic_stream_head_t *)picosplay_last(&cnx->stream_tree);
-#endif
-}
-
-int picoquic_compare_stream_priority(picoquic_stream_head_t * stream, picoquic_stream_head_t * other) {
-    int ret = 1;
-    if (stream->stream_priority < other->stream_priority) {
-        ret = -1;
-    }
-    else if (stream->stream_priority == other->stream_priority) {
-        if (stream->stream_id < other->stream_id) {
-            ret = -1;
-        }
-        else if (stream->stream_id == other->stream_id) {
-            ret = 0;
-        }
-    }
-    return ret;
-}
-
-/* This code assumes that the stream is not currently present in the output stream.
- */
-void picoquic_insert_output_stream(picoquic_cnx_t* cnx, picoquic_stream_head_t* stream)
-{
-    if (stream->is_output_stream == 0)  
-    {
-        if (IS_CLIENT_STREAM_ID(stream->stream_id) == cnx->client_mode) {
-            if (stream->stream_id > ((IS_BIDIR_STREAM_ID(stream->stream_id)) ? cnx->max_stream_id_bidir_remote : cnx->max_stream_id_unidir_remote)) {
-                return;
-            }
-        }
-
-        if (cnx->last_output_stream == NULL) {
-            /* insert first stream */
-            cnx->last_output_stream = stream;
-            cnx->first_output_stream = stream;
-        }
-        else if (picoquic_compare_stream_priority(stream, cnx->last_output_stream) >= 0) {
-            /* insert after last stream. Common case for most applications. */
-            stream->previous_output_stream = cnx->last_output_stream;
-            cnx->last_output_stream->next_output_stream = stream;
-            cnx->last_output_stream = stream;
-        }
-        else {
-            picoquic_stream_head_t* current = cnx->first_output_stream;
-
-            while (current != NULL) {
-                int cmp = picoquic_compare_stream_priority(stream, current);
-
-                if (cmp < 0) {
-                    /* insert before the current stream, then break */
-                    stream->previous_output_stream = current->previous_output_stream;
-                    if (stream->previous_output_stream == NULL) {
-                        cnx->first_output_stream = stream;
-                    }
-                    else {
-                        stream->previous_output_stream->next_output_stream = stream;
-                    }
-                    current->previous_output_stream = stream;
-                    stream->next_output_stream = current;
-                    break;
-                }
-                else if (cmp == 0) {
-                    /* Stream is already there. This is unexpected */
-                    break;
-                }
-                else {
-                    current = current->next_output_stream;
-                }
-            }
-            if (current == NULL) {
-                /* insert after last stream */
-                stream->previous_output_stream = cnx->last_output_stream;
-                cnx->last_output_stream->next_output_stream = stream;
-                cnx->last_output_stream = stream;
-            }
-        }
-
-        stream->is_output_stream = 1;
-    }
-}
-
-void picoquic_remove_output_stream(picoquic_cnx_t* cnx, picoquic_stream_head_t * stream)
-{
-    if (stream->is_output_stream) {
-        stream->is_output_stream = 0;
-
-        if (stream->previous_output_stream == NULL) {
-            cnx->first_output_stream = stream->next_output_stream;
-        }
-        else {
-            stream->previous_output_stream->next_output_stream = stream->next_output_stream;
-        }
-
-        if (stream->next_output_stream == NULL) {
-            cnx->last_output_stream = stream->previous_output_stream;
-        }
-        else {
-            stream->next_output_stream->previous_output_stream = stream->previous_output_stream;
-        }
-        stream->previous_output_stream = NULL;
-        stream->next_output_stream = NULL;
-    }
-}
-
-/* Reorder streams by priorities and rank.
- * A stream is deemed out of order if:
- * - the previous stream in the list has a higher priority, or
- * - the new stream has a lower priority.
- */
-void picoquic_reorder_output_stream(picoquic_cnx_t* cnx, picoquic_stream_head_t* stream)
-{
-    if (stream->is_output_stream) {
-        if ((stream->previous_output_stream != NULL &&
-            picoquic_compare_stream_priority(stream, stream->previous_output_stream) < 0) ||
-            (stream->next_output_stream != NULL &&
-                picoquic_compare_stream_priority(stream, stream->next_output_stream) > 0)) {
-            picoquic_remove_output_stream(cnx, stream);
-            stream->is_output_stream = 0;
-            picoquic_insert_output_stream(cnx, stream);
-        }
-    }
-}
-
-picoquic_stream_head_t * picoquic_next_stream(picoquic_stream_head_t * stream)
-{
-    return (picoquic_stream_head_t *)picosplay_next((picosplay_node_t *)stream);
-}
-
-picoquic_stream_head_t* picoquic_find_stream(picoquic_cnx_t* cnx, uint64_t stream_id)
-{
-    picoquic_stream_head_t target;
-    target.stream_id = stream_id;
-
-    return (picoquic_stream_head_t *)picosplay_find(&cnx->stream_tree, (void*)&target);
-}
-
-void picoquic_add_output_streams(picoquic_cnx_t* cnx, uint64_t old_limit, uint64_t new_limit, unsigned int is_bidir)
-{
-    uint64_t old_rank = STREAM_RANK_FROM_ID(old_limit);
-    uint64_t first_new_id = STREAM_ID_FROM_RANK(old_rank + 1ull, cnx->client_mode, !is_bidir);
-    picoquic_stream_head_t* stream = picoquic_find_stream(cnx, first_new_id );
-
-    while (stream) {
-        if (stream->stream_id > old_limit) {
-            if (stream->stream_id > new_limit) {
-                break;
-            }
-            if (IS_LOCAL_STREAM_ID(stream->stream_id, cnx->client_mode) && IS_BIDIR_STREAM_ID(stream->stream_id) == is_bidir) {
-                picoquic_insert_output_stream(cnx, stream);
-            }
-        }
-        stream = picoquic_next_stream(stream);
-    }
-}
-
-picoquic_stream_head_t* picoquic_create_stream(picoquic_cnx_t* cnx, uint64_t stream_id)
-{
-    picoquic_stream_head_t* stream = (picoquic_stream_head_t*)malloc(sizeof(picoquic_stream_head_t));
-    if (stream != NULL) {
-        memset(stream, 0, sizeof(picoquic_stream_head_t));
-        picoquic_sack_list_init(&stream->sack_list);
-    }
-
-    if (stream != NULL){
-        int is_output_stream = 0;
-        stream->stream_id = stream_id;
-        stream->cnx = cnx;
-
-        if (IS_LOCAL_STREAM_ID(stream_id, cnx->client_mode)) {
-            if (IS_BIDIR_STREAM_ID(stream_id)) {
-                stream->maxdata_local = cnx->local_parameters.initial_max_stream_data_bidi_local;
-                stream->maxdata_remote = cnx->remote_parameters.initial_max_stream_data_bidi_remote;
-                is_output_stream = stream->stream_id <= cnx->max_stream_id_bidir_remote;
-
-            }
-            else {
-                stream->maxdata_local = 0;
-                stream->maxdata_remote = cnx->remote_parameters.initial_max_stream_data_uni;
-                is_output_stream = stream->stream_id <= cnx->max_stream_id_unidir_remote;
-            }
-        }
-        else {
-            if (IS_BIDIR_STREAM_ID(stream_id)) {
-                stream->maxdata_local = cnx->local_parameters.initial_max_stream_data_bidi_remote;
-                stream->maxdata_remote = cnx->remote_parameters.initial_max_stream_data_bidi_local;
-                is_output_stream = 1;
-            }
-            else {
-                stream->maxdata_local = cnx->local_parameters.initial_max_stream_data_uni;
-                stream->maxdata_remote = 0;
-                is_output_stream = 0;
-            }
-        }
-
-        stream->stream_priority = cnx->quic->default_stream_priority;
-
-        picosplay_init_tree(&stream->stream_data_tree, picoquic_stream_data_node_compare, picoquic_stream_data_node_create, picoquic_stream_data_node_delete, picoquic_stream_data_node_value);
-
-        picosplay_insert(&cnx->stream_tree, stream);
-        if (is_output_stream) {
-            picoquic_insert_output_stream(cnx, stream);
-        }
-        else {
-            picoquic_remove_output_stream(cnx, stream);
-            picoquic_delete_stream_if_closed(cnx, stream);
-        }
-
-        if (stream_id >= cnx->next_stream_id[STREAM_TYPE_FROM_ID(stream_id)]) {
-            cnx->next_stream_id[STREAM_TYPE_FROM_ID(stream_id)] = NEXT_STREAM_ID_FOR_TYPE(stream_id);
-        }
-    }
-
-    return stream;
-}
-
-void picoquic_delete_stream(picoquic_cnx_t * cnx, picoquic_stream_head_t* stream)
-{
-    picosplay_delete(&cnx->stream_tree, stream);
-}
-
-int picoquic_mark_direct_receive_stream(picoquic_cnx_t* cnx, uint64_t stream_id, picoquic_stream_direct_receive_fn direct_receive_fn, void* direct_receive_ctx)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream;
-    picoquic_stream_data_node_t* data;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    if ((stream = picoquic_find_stream(cnx, stream_id)) == NULL) {
-        ret = PICOQUIC_ERROR_INVALID_STREAM_ID;
-    }
-    else if (!IS_BIDIR_STREAM_ID(stream_id) && IS_LOCAL_STREAM_ID(stream_id, cnx->client_mode)) {
-        ret = PICOQUIC_ERROR_INVALID_STREAM_ID;
-    }
-    else if (direct_receive_fn == NULL) {
-        /* This is illegal! */
-        ret = PICOQUIC_ERROR_NO_CALLBACK_PROVIDED;
-    }
-    else {
-        stream->direct_receive_fn = direct_receive_fn;
-        stream->direct_receive_ctx = direct_receive_ctx;
-        /* If there is pending data, pass it. */
-        while ((data = (picoquic_stream_data_node_t*)picosplay_first(&stream->stream_data_tree)) != NULL) {
-            size_t length = data->length;
-            uint64_t offset = data->offset;
-
-            if (offset < stream->consumed_offset) {
-                if (offset + length < stream->consumed_offset) {
-                    length = 0;
-                }
-                else {
-                    size_t delta_offset = (size_t)(stream->consumed_offset - offset);
-                    length -= delta_offset;
-                    offset += delta_offset;
-                }
-            }
-
-            if (length > 0) {
-                ret = direct_receive_fn(cnx, stream_id, 0, data->bytes, offset, length, direct_receive_ctx);
-            }
-
-            if (ret == 0) {
-                picosplay_delete_hint(&stream->stream_data_tree, &data->stream_data_node);
-            }
-            else {
-                break;
-            }
-        }
-
-        /* If there is a fin offset, pass it. */
-        if (ret == 0 && stream->fin_received && !stream->fin_signalled) {
-            uint8_t fin_bytes[8];
-            ret = direct_receive_fn(cnx, stream_id, 1, fin_bytes, stream->fin_offset, 0, direct_receive_ctx);
-        }
-    }
-
-    return ret;
-}
 
 
 /* Management of local CID.
@@ -4349,7 +3928,7 @@ picoquic_cnx_t* picoquic_create_cnx_internal(picoquic_quic_t* quic,
             cnx->tls_stream[epoch].maxdata_local = UINT64_MAX;
             cnx->tls_stream[epoch].maxdata_remote = UINT64_MAX;
 
-            picosplay_init_tree(&cnx->tls_stream[epoch].stream_data_tree, picoquic_stream_data_node_compare, picoquic_stream_data_node_create, picoquic_stream_data_node_delete, picoquic_stream_data_node_value);
+            picoquic_init_tls_tree(cnx, epoch);
             picoquic_sack_list_init(&cnx->tls_stream[epoch].sack_list);
             /* No need to reset the state flags, as they are not used for the crypto stream */
         }
@@ -4368,7 +3947,7 @@ picoquic_cnx_t* picoquic_create_cnx_internal(picoquic_quic_t* quic,
         cnx->min_ack_delay_local = cnx->ack_frequency_delay_local;
 
 
-        picosplay_init_tree(&cnx->stream_tree, picoquic_stream_node_compare, picoquic_stream_node_create, picoquic_stream_node_delete, picoquic_stream_node_value);
+        picoquic_init_stream_tree(cnx);
 
         cnx->congestion_alg = cnx->quic->default_congestion_alg;
         cnx->congestion_alg_option_string = cnx->quic->default_congestion_alg_option_string;
