+++ picoquic/sender.c
@@ -47,62 +47,6 @@
  * generate new packets, which are queued in the chained list.
  */
 
-static picoquic_stream_head_t* picoquic_find_stream_for_writing(picoquic_cnx_t* cnx,
-    uint64_t stream_id, int * ret)
-{
-    picoquic_stream_head_t* stream = picoquic_find_stream(cnx, stream_id);
-
-    *ret = 0;
-
-    if (stream == NULL) {
-        /* Need to check that the ID is authorized */
-
-        /* Check parity */
-        if (IS_CLIENT_STREAM_ID(stream_id) != cnx->client_mode) {
-            *ret = PICOQUIC_ERROR_INVALID_STREAM_ID;
-        }
-
-        if (*ret == 0) {
-            if (stream_id < cnx->next_stream_id[STREAM_TYPE_FROM_ID(stream_id)]) {
-                *ret = PICOQUIC_ERROR_STREAM_ALREADY_CLOSED;
-            } else {
-                stream = picoquic_create_missing_streams(cnx, stream_id, 0);
-                if (stream == NULL) {
-                    *ret = PICOQUIC_ERROR_MEMORY;
-                }
-            }
-        }
-    }
-
-    return stream;
-}
-
-int picoquic_set_app_stream_ctx(picoquic_cnx_t* cnx,
-    uint64_t stream_id, void* app_stream_ctx)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-    
-    stream = picoquic_find_stream_for_writing(cnx, stream_id, &ret);
-    if (ret == 0) {
-        stream->app_stream_ctx = app_stream_ctx;
-    }
-
-    return ret;
-}
-
-void picoquic_unlink_app_stream_ctx(picoquic_cnx_t* cnx, uint64_t stream_id)
-{
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic); 
-    
-    stream = picoquic_find_stream(cnx, stream_id);
-    if (stream != NULL) {
-        stream->app_stream_ctx = NULL;
-    }
-}
-
 int picoquic_mark_datagram_ready(picoquic_cnx_t* cnx, int is_ready)
 {
     int ret = 0;
@@ -143,55 +87,6 @@ int picoquic_mark_datagram_ready_path(picoquic_cnx_t* cnx, uint64_t unique_path_
     return ret;
 }
 
-
-int picoquic_mark_active_stream(picoquic_cnx_t* cnx,
-    uint64_t stream_id, int is_active, void * app_stream_ctx)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    stream = picoquic_find_stream_for_writing(cnx, stream_id, &ret);
-    if (ret == 0) {
-        if (is_active) {
-            /* The call only fails if the stream was closed or reset */
-            if (!stream->fin_requested && 
-                (!stream->reset_requested || picoquic_check_sack_list(&stream->sack_list, 0, stream->reliable_size) == 0) &&
-                cnx->callback_fn != NULL) {
-                stream->app_stream_ctx = app_stream_ctx;
-                if (!stream->is_active) {
-                    stream->is_active = 1;
-                    picoquic_reinsert_by_wake_time(cnx->quic, cnx, picoquic_get_quic_time(cnx->quic));
-                }
-            }
-            else {
-                ret = PICOQUIC_ERROR_CANNOT_SET_ACTIVE_STREAM;
-            }
-        }
-        else {
-            stream->is_active = 0;
-            stream->app_stream_ctx = app_stream_ctx;
-        }
-    }
-
-    return ret;
-}
-
-int picoquic_set_stream_not_coalesced(picoquic_cnx_t* cnx, uint64_t stream_id, int is_not_coalesced)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    stream = picoquic_find_stream_for_writing(cnx, stream_id, &ret);
-    if (ret == 0) {
-        stream->is_not_coalesced = is_not_coalesced;
-    }
-
-    return ret;
-}
-
-
 void picoquic_set_default_datagram_priority(picoquic_quic_t* quic, uint8_t default_datagram_priority)
 {
     PICOQUIC_THREAD_CHECK(quic);
@@ -202,118 +97,6 @@ void picoquic_set_datagram_priority(picoquic_cnx_t* cnx, uint8_t datagram_priori
 {
     PICOQUIC_THREAD_CHECK(cnx->quic);
     cnx->datagram_priority = datagram_priority;
-}
-
-void picoquic_set_default_priority(picoquic_quic_t* quic, uint8_t default_stream_priority)
-{
-    PICOQUIC_THREAD_CHECK(quic);
-    quic->default_stream_priority = default_stream_priority;
-}
-
-int picoquic_set_stream_priority(picoquic_cnx_t* cnx, uint64_t stream_id, uint8_t stream_priority)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    stream = picoquic_find_stream_for_writing(cnx, stream_id, &ret);
-    if (ret == 0) {
-        stream->stream_priority = stream_priority;
-        picoquic_reorder_output_stream(cnx, stream);
-    }
-
-    return ret;
-}
-
-int picoquic_mark_high_priority_stream(picoquic_cnx_t * cnx, uint64_t stream_id, int is_high_priority)
-{
-    int ret;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    if (is_high_priority) {
-        cnx->high_priority_stream_id = stream_id;
-    }
-    else if (cnx->high_priority_stream_id == stream_id) {
-        cnx->high_priority_stream_id = UINT64_MAX;
-    }
-
-    ret = picoquic_set_stream_priority(cnx, stream_id, (is_high_priority) ? 0 : cnx->quic->default_stream_priority);
-
-    return ret;
-}
-
-int picoquic_add_to_stream_with_ctx(picoquic_cnx_t* cnx, uint64_t stream_id,
-    const uint8_t* data, size_t length, int set_fin, void * app_stream_ctx)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-        
-    stream = picoquic_find_stream_for_writing(cnx, stream_id, &ret);
-    if (ret == 0 && set_fin) {
-        if (stream->fin_requested) {
-            /* app error, notified the fin twice*/
-            if (length > 0) {
-                ret = -1;
-            }
-        } else {
-            stream->fin_requested = 1;
-        }
-    }
-
-    /* If our side has sent RST_STREAM or received STOP_SENDING, we should not send anymore data. */
-    if (ret == 0 && (stream->reset_sent || stream->stop_sending_received)) {
-        ret = -1;
-    }
-
-    if (ret == 0 && length > 0) {
-        picoquic_stream_queue_node_t* stream_data = (picoquic_stream_queue_node_t*)
-            malloc(sizeof(picoquic_stream_queue_node_t));
-        if (stream_data == 0) {
-            ret = -1;
-        } else {
-            stream_data->bytes = (uint8_t*)malloc(length);
-
-            if (stream_data->bytes == NULL) {
-                free(stream_data);
-                stream_data = NULL;
-                ret = -1;
-            } else {
-                picoquic_stream_queue_node_t** pprevious = &stream->send_queue;
-                picoquic_stream_queue_node_t* next = stream->send_queue;
-
-                memcpy(stream_data->bytes, data, length);
-                stream_data->length = length;
-                stream_data->offset = 0;
-                stream_data->next_stream_data = NULL;
-
-                while (next != NULL) {
-                    pprevious = &next->next_stream_data;
-                    next = next->next_stream_data;
-                }
-
-                *pprevious = stream_data;
-            }
-        }
-
-    }
-
-    if (ret == 0) {
-        if (length > 0 || set_fin) {
-            picoquic_reinsert_by_wake_time(cnx->quic, cnx, picoquic_get_quic_time(cnx->quic));
-        }
-        cnx->nb_bytes_queued += length;
-        stream->is_active = 0;
-        stream->app_stream_ctx = app_stream_ctx;
-    }
-
-    return ret;
-}
-
-int picoquic_add_to_stream(picoquic_cnx_t* cnx, uint64_t stream_id,
-    const uint8_t* data, size_t length, int set_fin)
-{
-    return picoquic_add_to_stream_with_ctx(cnx, stream_id, data, length, set_fin, NULL);
 }
 
 int picoquic_set_app_flow_control(picoquic_cnx_t* cnx, uint64_t stream_id, int use_app_flow_control)
@@ -368,128 +151,6 @@ int picoquic_open_flow_control(picoquic_cnx_t* cnx, uint64_t stream_id, uint64_t
     return ret;
 }
 
-void picoquic_reset_stream_ctx(picoquic_cnx_t* cnx, uint64_t stream_id)
-{
-    picoquic_stream_head_t* stream;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    if ((stream = picoquic_find_stream(cnx, stream_id)) != NULL) {
-        stream->app_stream_ctx = NULL;
-    }
-}
-
-int picoquic_reset_stream_at(picoquic_cnx_t* cnx,
-    uint64_t stream_id, uint64_t local_stream_error, uint64_t reliable_size)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream = NULL;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    if (reliable_size > 0 && !cnx->is_reset_stream_at_enabled) {
-        ret = PICOQUIC_ERROR_ILLEGAL_TRANSPORT_EXTENSION;
-    }
-    else if ((stream = picoquic_find_stream(cnx, stream_id)) == NULL) {
-        ret = PICOQUIC_ERROR_INVALID_STREAM_ID;
-    }
-    else {
-        stream->app_stream_ctx = NULL;
-        if (stream->fin_sent && picoquic_check_sack_list(&stream->sack_list, 0, stream->fin_offset) == 0) {
-            ret = PICOQUIC_ERROR_STREAM_ALREADY_CLOSED;
-        }
-        else if (!stream->reset_requested) {
-            stream->local_error = local_stream_error;
-            stream->reset_requested = 1;
-            stream->reliable_size = reliable_size;
-        }
-    }
-
-    picoquic_reinsert_by_wake_time(cnx->quic, cnx, picoquic_get_quic_time(cnx->quic));
-
-    return ret;
-}
-int picoquic_reset_stream(picoquic_cnx_t* cnx,
-    uint64_t stream_id, uint64_t local_stream_error)
-{
-    return picoquic_reset_stream_at(cnx, stream_id, local_stream_error, 0);
-}
-
-uint64_t picoquic_get_next_local_stream_id(picoquic_cnx_t* cnx, int is_unidir)
-{
-    /* This code could be written as:
-     * int stream_type_id = ((cnx->client_mode ^ 1) | ((is_unidir) ? 2 : 0)); 
-     * but Visual Studio produces an obnoxious error message about
-     * mixing bitwise or and logical or. */
-    int stream_type_id = cnx->client_mode ^ 1;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    if (is_unidir) {
-        stream_type_id |= 2;
-    }
-
-    return cnx->next_stream_id[stream_type_id];     
-}
-
-int picoquic_stop_sending(picoquic_cnx_t* cnx,
-    uint64_t stream_id, uint64_t local_stream_error)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream = NULL;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    stream = picoquic_find_stream(cnx, stream_id);
-
-    if (stream == NULL) {
-        ret = PICOQUIC_ERROR_INVALID_STREAM_ID;
-    }
-    else {
-        stream->app_stream_ctx = NULL;
-
-        if (stream->reset_received) {
-            ret = PICOQUIC_ERROR_STREAM_ALREADY_CLOSED;
-        }
-        else if (!stream->stop_sending_requested) {
-            stream->local_stop_error = local_stream_error;
-            stream->stop_sending_requested = 1;
-            picoquic_insert_output_stream(cnx, stream);
-        }
-    }
-
-    picoquic_reinsert_by_wake_time(cnx->quic, cnx, picoquic_get_quic_time(cnx->quic));
-
-    return ret;
-}
-
-int picoquic_discard_stream(picoquic_cnx_t* cnx, uint64_t stream_id, uint16_t local_stream_error)
-{
-    int ret = 0;
-    picoquic_stream_head_t* stream = NULL;
-    PICOQUIC_THREAD_CHECK(cnx->quic);
-
-    stream = picoquic_find_stream(cnx, stream_id);
-
-    if (stream == NULL) {
-        ret = PICOQUIC_ERROR_INVALID_STREAM_ID;
-    }
-    else {
-        if (IS_BIDIR_STREAM_ID(stream_id) || !IS_CLIENT_STREAM_ID(stream_id)) {
-            ret = picoquic_stop_sending(cnx, stream_id, local_stream_error);
-            if (ret == PICOQUIC_ERROR_STREAM_ALREADY_CLOSED) {
-                ret = 0;
-            }
-        }
-        if (ret == 0 &&
-            (IS_BIDIR_STREAM_ID(stream_id) || IS_CLIENT_STREAM_ID(stream_id))) {
-            ret = picoquic_reset_stream(cnx, stream_id, local_stream_error);
-            if (ret == PICOQUIC_ERROR_STREAM_ALREADY_CLOSED) {
-                ret = 0;
-            }
-        }
-        stream->app_stream_ctx = NULL;
-        stream->is_discarded = 1;
-    }
-
-    return ret;
-}
 
 /*
  * Manage content padding
